+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v0.0.51 (02/08/2020)
+
+- Added support for creating playgrounds without backing gists
+- Added support for adding, renaming and deleting files within a temp playground
+- Added the `Delete File` command to the editor context menu
+
 ## v0.0.50 (02/07/2020)
 
 - Added a confirmation prompt when attempting to delete gist files